chore(lint) cleaning linting error
