+++ .eslintignore
@@ -1,1 +1,2 @@
 src/test
+src/lib/darknode/utils/mocks/
